Move contract info enrichment to service stub
The stub is actually just a function and its supporting types for now.
