+++ background/redux-slices/utils/activity-utils.ts
@@ -1,17 +1,7 @@
 import dayjs from "dayjs"
-import {
-  SmartContractFungibleAsset,
-  AnyAsset,
-  isSmartContractFungibleAsset,
-  AnyAssetAmount,
-} from "../../assets"
 import { convertToEth } from "../../lib/utils"
 import { AnyEVMTransaction } from "../../networks"
-import {
-  AssetDecimalAmount,
-  enrichAssetAmountWithDecimalValues,
-} from "./asset-utils"
-import { HexString } from "../../types"
+import { ContractInfo } from "../../services/enrichment"
 
 function ethTransformer(
   value: string | number | bigint | null | undefined
@@ -31,36 +21,6 @@ type FieldAdapter<T> = {
 export type UIAdaptationMap<T> = {
   [P in keyof T]?: FieldAdapter<T[P]>
 }
-
-export type BaseContractInfo = {
-  contractLogoURL?: string | undefined
-}
-
-export type ContractDeployment = BaseContractInfo & {
-  type: "contract-deployment"
-}
-
-export type ContractInteraction = BaseContractInfo & {
-  type: "contract-interaction"
-}
-
-export type AssetTransfer = BaseContractInfo & {
-  type: "asset-transfer"
-  assetAmount: AnyAssetAmount & AssetDecimalAmount
-}
-
-export type AssetSwap = BaseContractInfo & {
-  type: "asset-swap"
-  fromAssetAmount: AnyAssetAmount & AssetDecimalAmount
-  toAssetAmount: AnyAssetAmount & AssetDecimalAmount
-}
-
-export type ContractInfo =
-  | ContractDeployment
-  | ContractInteraction
-  | AssetTransfer
-  | AssetSwap
-  | undefined
 
 export type ActivityItem = AnyEVMTransaction & {
   contractInfo?: ContractInfo | undefined
@@ -168,78 +128,4 @@ export const keysMap: UIAdaptationMap<ActivityItem> = {
       return ""
     },
   },
-}
-
-function resolveContractInfo(
-  assets: AnyAsset[],
-  contractAddress: HexString | undefined,
-  contractInput: HexString,
-  desiredDecimals: number
-): ContractInfo | undefined {
-  // A missing recipient means a contract deployment.
-  if (typeof contractAddress === "undefined") {
-    return {
-      type: "contract-deployment",
-    }
-  }
-
-  // See if the address matches a fungible asset.
-  const matchingFungibleAsset = assets.find(
-    (asset): asset is SmartContractFungibleAsset =>
-      isSmartContractFungibleAsset(asset) &&
-      asset.contractAddress.toLowerCase() === contractAddress.toLowerCase()
-  )
-
-  const contractLogoURL = matchingFungibleAsset?.metadata?.logoURL
-
-  // Derive value from transaction transfer if not sending ETH
-  // FIXME Move to ERC20 parsing using ethers.
-  if (
-    typeof matchingFungibleAsset !== "undefined" &&
-    contractInput.length >= 74 &&
-    contractInput.startsWith("0xa9059cbb") // transfer selector
-  ) {
-    return {
-      type: "asset-transfer",
-      contractLogoURL,
-      assetAmount: enrichAssetAmountWithDecimalValues(
-        {
-          asset: matchingFungibleAsset,
-          amount: BigInt(`0x${contractInput.slice(10, 10 + 64)}`),
-        },
-        desiredDecimals
-      ),
-    }
-  }
-
-  // Fall back on a standard contract interaction.
-  return {
-    type: "contract-interaction",
-    contractLogoURL,
-  }
-}
-
-export function enrichTransactionWithContractInfo(
-  assets: AnyAsset[],
-  transaction: AnyEVMTransaction,
-  desiredDecimals: number
-): AnyEVMTransaction & { contractInfo?: ContractInfo | undefined } {
-  if (transaction.input === null || transaction.input === "0x") {
-    // This is _almost certainly_ not a contract interaction, move on. Note that
-    // a simple ETH send to a contract address can still effectively be a
-    // contract interaction (because it calls the fallback function on the
-    // contract), but for now we deliberately ignore that scenario when
-    // categorizing activities.
-    return transaction
-  }
-
-  return {
-    ...transaction,
-    contractInfo: resolveContractInfo(
-      assets,
-      transaction.to,
-      transaction.input,
-      desiredDecimals
-    ),
-  }
 }
